+++ tests/test_capability_list.py
@@ -29,7 +29,7 @@ class TestCapabilityList(unittest.TestCase):
     def test03_parse(self):
         xml='<?xml version=\'1.0\' encoding=\'UTF-8\'?>\n<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:rs="http://www.openarchives.org/rs/terms/"><rs:md capability="capabilitylist" from="2013-02-07T22:39:00" /><url><loc>http://example.org/resourcelist.xml</loc><rs:md capability="resourcelist" /></url></urlset>'
         cl=CapabilityList()
-        cl.parse(str=xml)
+        cl.parse(str_data=xml)
         self.assertEqual( cl.capability, 'capabilitylist')
         self.assertEqual( len(cl.resources), 1, 'got 1 resource')
         [r] = cl.resources