+++ tests/test_source_description.py
@@ -34,7 +34,7 @@ class TestSourceDescription(unittest.TestCase):
     def test04_parse(self):
         xml='<?xml version=\'1.0\' encoding=\'UTF-8\'?>\n<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:rs="http://www.openarchives.org/rs/terms/"><rs:ln href="http://example.org/about" rel="describedby" /><rs:md capability="description" /><url><loc>http://example.org/ds1/cl.xml</loc><rs:md capability="capabilitylist" /></url><url><loc>http://example.org/ds2/cl.xml</loc><rs:md capability="capabilitylist" /></url><url><loc>http://example.org/ds3/cl.xml</loc><rs:md capability="capabilitylist" /></url></urlset>'
         sd=SourceDescription()
-        sd.parse(str=xml)
+        sd.parse(str_data=xml)
         self.assertEqual( sd.link_href('describedby'), 'http://example.org/about', 
                           'describedby link' )
         self.assertEqual( sd.capability, 'description' )