+++ CHANGES.md
@@ -6,9 +6,11 @@ core specification version. Versions 1.0.x implement the v1.0
 ResourceSync specification which was standardized as ANSI/NISO Z39.99-2014
 <http://www.openarchives.org/rs/1.0/toc>.
 
-v1.0.1 2016-01-30
+v1.0.1 2016-03-22
   * Improved test coverage
   * Use defusedxml for safer XML parsing
+  * Changed parameter str to str_data in ListBase.parse(), str still supported
+    but deprecated. Issue pointed out by @uweschmitt
 
 v1.0.0 2014-06-18
   * Update for v1.0, Z39.99-2014 specification (http://www.openarchives.org/rs/1.0/)